add cast in testlassort

+++ src/test/testlassort.cpp
@@ -161,7 +161,7 @@ int main(int argc, char * argv[])
 			for ( uint64_t y = 0; y < numout; ++y )
 			{
 				uint64_t const ilow = y * fanin;
-				uint64_t const ihigh = std::min(ilow+fanin,V.size());
+				uint64_t const ihigh = std::min(ilow+fanin,static_cast<uint64_t>(V.size()));
 				uint64_t const isize = ihigh-ilow;
 
 				std::cerr << "[V] merging [" << ilow << "," << ihigh << ")" << std::endl;
